I confused myself trying to get the Giant stat multiplier to work.
It's about to turn out well. I just need a way to identify the correct buff when removing from the buff list.

+++ Assets/Scripts/Armour.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Armour : Equipment {
 
-    public Armour(string _name, string _description, Resources _resourceCost, Sprite _icon, int _craftingTime, Stats _stats, int _reinforcementCount, Stats _reinforcementStats)
+    public Armour(string _name, string _description, Resources _resourceCost, Sprite _icon, float _craftingTime, Stats _stats, int _reinforcementCount, Stats _reinforcementStats)
         : base(_name, _description, _resourceCost, _icon, _craftingTime, _stats, _reinforcementCount, _reinforcementStats)
     {
