+++ Assets/Scripts/Equipment.cs
@@ -7,7 +7,7 @@ public class Equipment : Item {
     public int reinforcementCount;
     public Stats reinforcementStats;
 
-    public Equipment(string _name, string _description, Resources _resourceCost, Sprite _icon, int _craftingTime, Stats _stats, int _reinforcementCount, Stats _reinforcementStats)
+    public Equipment(string _name, string _description, Resources _resourceCost, Sprite _icon, float _craftingTime, Stats _stats, int _reinforcementCount, Stats _reinforcementStats)
         : base(_name, _description, _resourceCost, _icon, _craftingTime)
     {
         stats = _stats;
